requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
efficientnet-pytorch==0.7.1
executing==2.0.1
filelock==3.13.0
fire==0.5.0
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.10.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
grpcio==1.59.0
idna==3.4
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
kiwisolver==1.4.5
lightning-utilities==0.9.0
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.1
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
python-dateutil==2.8.2
pytorch-lightning==2.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
thop==0.1.1.post2209072238
torch==2.1.0
torchmetrics==1.2.0
torchvision==0.16.0
tqdm==4.66.1
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.0.203
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==3.0.1
yarl==1.9.2
